refactor: use the term publish instead of create for feature
